examples/otlp_exporter_sample/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
grpcio-health-checking==1.59.3
grpcio-reflection==1.59.3
fastapi==0.110.0
uvicorn==0.29.0
gunicorn
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-logging